requirements.txt
grpcio==1.20.1
grpcio-tools==1.20.1
protobuf==3.7.1

retry==0.9.2
pymongo==3.8.0

coverage==4.5.3
flake8==3.7.7